Completed view and functioning of width of cloth (details)

+++ src/database/migrations/20220110194021-create-cloth.js
@@ -18,7 +18,7 @@ module.exports = {
         type: Sequelize.BOOLEAN
       },
       width:{
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(3,2)
       },
       idLocal :{
         type: Sequelize.INTEGER
